+++ src/openrct2/ride/Ride.h
@@ -306,7 +306,7 @@ struct Ride
     };
     uint8_t slide_peep_t_shirt_colour;
     uint8_t spiral_slide_progress;
-    uint16_t build_date;
+    int32_t build_date;
     money16 upkeep_cost;
     uint16_t race_winner;
     uint32_t music_position;
